feat: complete the pipeline with better error handling

+++ src/pipeline/docx.rs
@@ -1,9 +1,9 @@
 /// Convert cm to English metric unit
-pub fn cm(cm: f32) -> usize {
-    (cm * 360_000.0) as usize
+pub fn cm(cm: f32) -> u32 {
+    (cm * 360_000.0) as u32
 }
 
 /// Convert point to English metric unit
-pub fn pixel(point: f32) -> usize {
-    (point * 9525.0) as usize
+pub fn pixel(point: f32) -> u32 {
+    (point * 9525.0) as u32
 }
